34.4. Transition Name & Mode
The solution would be to make how we hide & show the components smarter. We would still rely on v-if to check for completeCart... because we can't render at all until that variable is ready. But then we would use v-show to check the current state. Thanks to that, instead of destroying the component, it would just hide it.

The problem is that... this doesn't work:

  <transition> can only be used on a single element. Use <transition-group> for lists, which is not our situation.

So basically... we can't wrap multiple things unless we're using v-if exclusively. The solution would be to wrap each component in its own transition tag: one around the shopping cart and another around the checkout form.
But when we do this, we can no longer use the mode option, which is what helped us fix that jumping problem. One common work-around is to absolutely position the two components on top of each other so that as they fade out and in, the do it in the same spot without the jumping.