34.5. Transition Name & Mode
Anyways, I'm going to undo this and go back to v-if: I just wanted to give you some extra background. I'll refresh to make sure I didn't break things... excellent!
There's one more thing that I want to transition: the title on the page. If you watch, it still changes instantly. That's not a huge deal, but it would be cool if that also faded out and faded back in. This is interesting because, in this case, we will be transitioning over a prop change. Let's talk about how to do that next.