35.2. Transitioning a Prop Change
Transitions and the key Prop
However, it is possible to, sort of, make this work. Over on the title component, add a new key prop. Yea, that's the prop we normally use in loops. Set this to something that will be unique for each title, like currentState, because we know the page title is based on the currentState.
Move over and try it now. Ah! It works! The title fades out and fades in just like the rest of the page!
So... how and why did that work? Especially right after I explained to you that you can't transition prop changes!
When we click the button, our top-level component re-renders and the title component now has a new value for its key prop. That actually tells Vue that the new title component should be a totally different instance than the one it was using before.
In other words, instead of re-rendering the same component instance... but passing in a new text prop, Vue completely destroys the old title component and creates a brand new one. This allows Vue to transition between the old and new title because the old component is hidden and the new component is shown.
This works great in our case... but it's kind of a "nuclear" option. If our title component had some data inside of it, that data would be completely lost each time we switched... because we're destroying and recreating the component. So it might work in some cases... but not in others.
For us, I'm super happy with this. So next: let's add a form to the checkout page!