36.1. Checkout Form
The goal of this page is to hold a checkout form. When the user submits that form, we will send an AJAX request to our API to create a purchase in the database, which will contain the customer's info from the form plus the list of the items in the user's cart.
To truly live our low-budget, minimum viable product hip startup mentality, instead of having a credit card form, after success, we'll redirect the user to a page with information about where to mail a check to. Very cutting edge.

Setting up the Data
Open assets/components/checkout/index.vue
This component will hold a bunch of form fields like customer name, email address, etc. And it's going to need to store those field values as data. Add a data function and return an object with a form key set to another object. Add a key for every field: customerName initialized to an empty string, customerEmail, customerAddress, customerZip, customerCity and customerPhone.
Next, because we are definitely going to add form validation, add a validationErrors data set to an empty object
This will eventually hold a map of which fields are currently invalid with their error message.

Adding the First Field
Nice! Up in the template, add the first field. Replace the text with a form tag, no action needed, then some Bootstrap markup: a div, the label - with "Name:" and a for attribute set to customerName. It's not necessary, but I've made this consistent with the data key... so I can keep my sanity. Add class="col-form-label" then add the <input and make sure its id matches the for label attribute. Also add v-model="form.customerName".
Thanks to this, whenever the user updates this input, v-model will also make sure that the form.customerName data is also updated. Finish with type=text and class=form-control.