36.2. Checkout Form
Good start! And if we check it in the browser... it's there! Inspect element, open the Vue dev tools and find the CheckoutForm component. We can watch the form.customerName data stay in sync thanks to v-model.
Oh, but to be extra cool, we can use v-model.trim, which is a shortcut to trim off any extra whitespace.
Now if I go back... and put a bunch of spaces at the start, the data does not have those.