36.3. Checkout Form
Rendering Validation Errors
We don't have any form validation logic yet, but let's prepare our field to be able to render the validation message based on the validationErrors data. To help, add a new methods section at the bottom of the component with one function: isFieldValid() with a fieldName argument. Inside, return not fieldName in this.validationErrors.
Use this above. First, on the input, if validation fails, this needs an extra class. Change to use :class and set form-control to true so that it always renders. We also want an is-invalid class if not isFieldValid('customerName').
Then, after the input, add a span with v-show set to that same thing: not isFieldValid('customerName') so that this only shows when the field is invalid. Give this class="invalid-feedback" and print the validation error inside: validationErrors.customerName.
And... hmm... ah! ESLint is mad because I messed up my indentation. Good work ESLint!

Vue Router?
Let's try this! Back at the browser, the page already refreshed... which highlights one annoying thing about our setup: when we refresh, it takes us back to the cart page! This is because we are not changing the URL when we go back and forth between the cart and checkout. And so, our component can't remember that we're on a different page.
Depending on your situation, this might be ok... or it might totally not be ok! This is something that could be improved by using the Vue Router: a topic we'll talk about in a future tutorial.
Anyways, click back to the checkout form, find the Vue dev tools, locate the CheckoutForm component, edit the validationErrors data and set it to an object with a customerName key and some message. Hit save. That looks lovely!
So... our customerName field is ready! There's just one problem: do we really want to copy and paste all of this 5 more times for the other 5 fields? Um... no. We can do better. Next, let's create a new form input component that we can re-use for every field.