38.1. v-model on a Custom Component
Our new <form-input/> mostly works and we'll be able to reuse it to render all of our fields. The only missing piece is that when we type into the field, it doesn't update the data in CheckoutForm.
Normally we would fix this by emitting an event from the child component so that we can listen to that event in the parent component and update the data that lives there. So let's do that! And... then do something cooler.

Emitting and Listening to the Event
In form-input.vue, on the <input/> element, add @input="" and I'll use the inline format to emit an event. Let's call it input to match the name of the event that's used for a normal form element. For the data, let's just pass the new value $event.target.value.
Cool! Now, in the parent component, we can listen to this: @input="" and then immediately update the data with form.customerName = $event.
Alrighty! Let's try it! Move over, click to check out, type in a name... and look at the data. Perfect! It is updating.