39.4. v-bind Many Props
Customizing the input type Attribute
Let's make one last improvement. All of these fields are <input type="text">. If we wanted to handle other field types like select elements or checkboxes, we would need to do more work in <form-input> to make it more flexible or even create some new components.
I'm not going to do that now, but I at least want to be able to render different input types, like <input type="email"> and <input type="tel"> for the phone number.
No problem... then scroll up so we can use it. Let's see: make the aside not be self-closing... then put <cart-sidebar inside with a v-if set to featuredProduct. Our <form-input> now needs to be more flexible. So let's add a new prop. Copy the value prop, call this one type... and change the default to text so that we don't have to pass this in.
Use this up in the template: replace type="text" with :type="type".
Thanks to the default value, we only need to pass this for two fields. Find customerEmail. What's cool is that we can mix the v-bind that's set to an entire object with other, specific props without any issues. What I mean is, when we pass in the type prop with type="email", that will merge nicely with whatever props getFieldProps() adds.
Repeat this on customerPhone: type="tel".
Go check it! You probably won't notice any difference on a computer, but if you inspect the element... yep! It's <input type="email">.
Okay! We are ready to set up this form to submit via Ajax. When we do that, we're going to make handling and rendering form validation a main concern.