40.2. Checkout Submit
We're also going to need a loading animation while this is saving. So down in data, add a new key: loading: false.
Then pull in our trusty loading component: import Loading from @/components/loading. Add that to components and up in the template, inside that new div, say <loading/> with v-show="loading".

Sending the AJAX Call
Our next step is pretty straightforward: we need to add a method that, when the form submits, sends the Ajax request. Down under methods add a new one called async onSubmit(). I'm immediately making this async because I know I'm going to want to await for the AJAX call to finish. Start with this.loading = true.
Because this is a checkout form, I don't want anything weird to happen. And so, I'm going to write more error handling than we've done so far. Add a try {} catch and a finally. Inside that last section, set this.loading = false.
Thanks to this - whether we're successful or not - the loading animation will hide.
Inside try, say const response = await createOrder(). Make sure you hit tab so that PhpStorm adds the import on top.
For the data argument, we need to pass the individual form fields and an extra purchaseItems field. The form fields are stored on the form data. So what we can do down here is say ...this.form. That will expand the object so that the data will have those 6 keys.
The last field we need to send is purchaseItems. For now, set that to an empty array.
Down in catch, use console.error() - that's like console.log()... but will look red - to log error.response. In Axios, if an AJAX call fails, it stores the error response on this key. Oh, and also console.log() the success response up in try.... actually response.data.
Awesome! We're eventually going to do something on success and error, but this is a good start.