43.1. Client-Side Validation
Click to go back to view the shopping cart... which is now empty. We're not done shopping! Add another inflatable sofa. Awesome!
Server-side validation is a must: we can't have API endpoints that allow any data. And we've handled that nicely. Go us!
Client-side validation in JavaScript is optional. It requires extra work, but it can make your form even nicer to use. So let's add some! We'll keep it simple: on "blur" of a field - so when we leave a field - if that field is empty, we'll immediately show a validation error.
Start inside of our checkout form component... down in methods. Add a new one called validateField(). I'll paste in the start: an object that contains the validation messages that should be used for each field if that field is blank.

Listening to blur on a Custom Component
Before we add the real logic to this method, let's call this "on blur" of each field. Easy! Head up to the template and add @blur="validateField".
Except... bah! That won't work! This is a custom component, not a normal form element.
Okay, don't panic. Go into form-input.vue. And on the actual <input> element, add @blur=""... and then we'll just emit that same blur event from here... and pass it the same event data.
We're basically propagating that event up so that our parent component can listen to it.
Back in index.vue, now that this should work, copy the @blur and repeat this on all six fields... super fast!