43.2. Client-Side Validation
Filling in the Validation Logic
Ok! Head back down to validateField()... here it is. The plan is: read the id attribute of whatever input was just blurred, use that to see if the form data for that field is empty and, if it is, add a new item to the validationErrors data.
To read the id attribute, we need the event argument. Then, down here, we can say const validationField = event.target.id so that will be something like customerName or customerEmail. Then, if not this.form[validationField], then we know that field is empty. Add a validation error with this.validationErrors[validationField] = validationMessages[validationField].
Oh, and don't forget an else: in case the user just went back... filled something in... and then blurred. If there was already an error before, now we need to remove it. Do that with delete this.validationErrors[validationField].
The reason I'm using delete and not equals null is... well... that's just the way that our validationErrors object works right now. It starts as an empty object and we only put things onto it when they have an error. We even reset it back to an empty object on submit. This will actually cause a "reactivity" problem... which we'll explore in a few minutes.
By the way, there are some Vue libraries to help with validation, like Vuelidate. You can totally check these out. They make adding a lot of complex validation rules a lot easier and cleaner. Right now, we're just checking to see if the field is blank. But even with something like Vuelidate, you're still basically doing the same thing: you're storing the errors for a field somewhere like validationErrors and then using that when you render to display the error.
Anyways, let's test this thing! Find your browser, click "Check Out", click into the name field... and hit tab. Huh. Nothing happened. Go check out the Vue dev tools, find CheckoutForm... then down to validationErrors. Yea, it says that we do have two validation errors for name and email: the two fields we blurred.
So the data is right.... but they're not rendering on the form. Want to see something even weirder? Go into one of the other boxes and type something. Boom! As soon as we do, both error messages show up! We don't see any errors until we modify a field.
What the heck is going on? The short answer is that we accidentally removed reactivity from the validationErrors data. That's a fancy way of saying that when we change a value on validationErrors, Vue does not realize that it needs to re-render. So the data is being stored correctly... but it doesn't reflect on the page until Vue re-renders for some other reason, like us changing some of the form data.
Next: let's explore this deeper... and fix it!